Translate tocsym.c to tocsym.d

+++ src/backend/symbol.c
@@ -267,7 +267,7 @@ symbol * symbol_calloc(const char *id)
     return symbol_calloc(id, strlen(id));
 }
 
-symbol * symbol_calloc(const char *id, size_t len)
+symbol * symbol_calloc(const char *id, unsigned len)
 {   symbol *s;
 
     //printf("sizeof(symbol)=%d, sizeof(s->Sident)=%d, len=%d\n",sizeof(symbol),sizeof(s->Sident),(int)len);
@@ -297,7 +297,7 @@ symbol * symbol_name(const char *name,int sclass,type *t)
     return symbol_name(name, strlen(name), sclass, t);
 }
 
-symbol * symbol_name(const char *name, size_t len, int sclass, type *t)
+symbol * symbol_name(const char *name, unsigned len, int sclass, type *t)
 {
     type_debug(t);
     symbol *s = symbol_calloc(name, len);
